requirements.txt
click==8.1.3
dash==2.8.1
dash-bootstrap-components==1.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
pandas==1.5.3
plotly==5.13.0
pyodbc==4.0.34
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.2.1
typing_extensions==4.4.0
Werkzeug==2.2.2
